Refactor: handle_options()
In order to do this refactor, I tried to minimize the complexity of if
and else if statements, by calling a function that handle each
conditional that handle commands. This functions are stored in
a CMD_STRATEGIES struct which encapsulates a family of algorithms for each case.

Signed-off-by thiago16fca <thiago16fca@outlook.com>